requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pyjwt==2.9.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1